refractor: fix pdfgeneration imports & file openblobs -> fileActionUtil.ts
